feed: fix incorrect LBPE with miss-matched slot count

+++ src/entity/S2Profile.ts
@@ -38,7 +38,7 @@ export class S2Profile {
     name: string;
 
     @Column({
-        type: 'smallint',
+        type: 'mediumint',
         unsigned: true,
     })
     discriminator: number;
